+++ vscode/src/completions/providers/createProvider.test.ts
@@ -85,7 +85,7 @@ describe('createProviderConfig', () => {
         it('returns "fireworks" provider config and corresponding model if specified', async () => {
             const provider = await createProviderConfig(
                 getVSCodeSettings({
-                    autocompleteAdvancedProvider: 'unstable-fireworks',
+                    autocompleteAdvancedProvider: 'fireworks',
                     autocompleteAdvancedModel: 'starcoder-3b',
                 }),
                 dummyCodeCompletionsClient,
@@ -97,7 +97,7 @@ describe('createProviderConfig', () => {
 
         it('returns "fireworks" provider config if specified in settings and default model', async () => {
             const provider = await createProviderConfig(
-                getVSCodeSettings({ autocompleteAdvancedProvider: 'unstable-fireworks' }),
+                getVSCodeSettings({ autocompleteAdvancedProvider: 'fireworks' }),
                 dummyCodeCompletionsClient,
                 {}
             )