+++ vscode/src/completions/providers/createProvider.ts
@@ -6,11 +6,8 @@ import { logError } from '../../log'
 import { CodeCompletionsClient } from '../client'
 
 import { createProviderConfig as createAnthropicProviderConfig } from './anthropic'
+import { createProviderConfig as createFireworksProviderConfig, FireworksOptions } from './fireworks'
 import { ProviderConfig } from './provider'
-import {
-    createProviderConfig as createUnstableFireworksProviderConfig,
-    UnstableFireworksOptions,
-} from './unstable-fireworks'
 import { createProviderConfig as createUnstableOpenAIProviderConfig } from './unstable-openai'
 
 export async function createProviderConfig(
@@ -25,7 +22,7 @@ export async function createProviderConfig(
         config.autocompleteAdvancedProvider
     )
     if (providerAndModelFromVSCodeConfig) {
-        const { provider, model } = providerAndModelFromVSCodeConfig
+        const { provider, model, starcoderExtendedTokenWindow } = providerAndModelFromVSCodeConfig
 
         switch (provider) {
             case 'unstable-openai': {
@@ -33,10 +30,11 @@ export async function createProviderConfig(
                     client,
                 })
             }
-            case 'unstable-fireworks': {
-                return createUnstableFireworksProviderConfig({
+            case 'fireworks': {
+                return createFireworksProviderConfig({
                     client,
                     model: config.autocompleteAdvancedModel ?? model ?? null,
+                    starcoderExtendedTokenWindow,
                 })
             }
             case 'anthropic': {
@@ -82,7 +80,7 @@ export async function createProviderConfig(
                 })
 
             case 'fireworks':
-                return createUnstableFireworksProviderConfig({
+                return createFireworksProviderConfig({
                     client,
                     model: model ?? null,
                 })
@@ -108,20 +106,24 @@ export async function createProviderConfig(
     })
 }
 
-async function resolveDefaultProviderFromVSCodeConfigOrFeatureFlags(
-    configuredProvider: string | null
-): Promise<{ provider: string; model?: UnstableFireworksOptions['model'] } | null> {
+async function resolveDefaultProviderFromVSCodeConfigOrFeatureFlags(configuredProvider: string | null): Promise<{
+    provider: string
+    model?: FireworksOptions['model']
+    starcoderExtendedTokenWindow?: boolean
+} | null> {
     if (configuredProvider) {
         return { provider: configuredProvider }
     }
 
-    const [starCoder7b, starCoder16b, starCoderHybrid, llamaCode7b, llamaCode13b] = await Promise.all([
-        featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoder7B),
-        featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoder16B),
-        featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoderHybrid),
-        featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteLlamaCode7B),
-        featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteLlamaCode13B),
-    ])
+    const [starCoder7b, starCoder16b, starCoderHybrid, llamaCode7b, llamaCode13b, starcoderExtendedTokenWindow] =
+        await Promise.all([
+            featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoder7B),
+            featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoder16B),
+            featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoderHybrid),
+            featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteLlamaCode7B),
+            featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteLlamaCode13B),
+            featureFlagProvider.evaluateFeatureFlag(FeatureFlag.CodyAutocompleteStarCoderExtendedTokenWindow),
+        ])
 
     if (starCoder7b || starCoder16b || starCoderHybrid || llamaCode7b || llamaCode13b) {
         const model = starCoder7b
@@ -133,7 +135,7 @@ async function resolveDefaultProviderFromVSCodeConfigOrFeatureFlags(
             : llamaCode7b
             ? 'llama-code-7b'
             : 'llama-code-13b'
-        return { provider: 'unstable-fireworks', model }
+        return { provider: 'fireworks', model, starcoderExtendedTokenWindow }
     }
 
     return null